+++ rr/ns.js
@@ -14,7 +14,7 @@ class NS extends RR {
     this.isFullyQualified('NS', 'dname', val)
     this.isValidHostname('NS', 'dname', val)
 
-    this.set('dname', val)
+    this.set('dname', val.toLowerCase())
   }
 
   getDescription () {
@@ -42,7 +42,7 @@ class NS extends RR {
     return new this.constructor({
       type     : 'NS',
       name     : this.fullyQualify(fqdn),
-      dname    : this.fullyQualify(dname),
+      dname    : this.fullyQualify(/\./.test(dname) ? dname : `${dname}.ns.${fqdn}`),
       ttl      : parseInt(ttl, 10),
       timestamp: ts,
       location : loc !== '' && loc !== '\n' ? loc : '',
@@ -63,6 +63,10 @@ class NS extends RR {
   }
 
   /******  EXPORTERS   *******/
+  toBind (zone_opts) {
+    return `${this.getPrefix(zone_opts)}\t${this.getFQDN('dname', zone_opts)}\n`
+  }
+
   toTinydns () {
     return `&${this.getTinyFQDN('name')}::${this.getTinyFQDN('dname')}:${this.getTinydnsPostamble()}\n`
   }